HotFix v3.2.0 add azure option azure-domain-hint
Add option domain_hint for azure provider,

domain_hint is optional.    Provides a hint about the tenant or domain
that the user should use to sign in. The value of the domain_hint is a
registered domain for the tenant. If the tenant is federated to an
on-premises directory, AAD redirects to the specified tenant federation
server.

Ref :
https://docs.microsoft.com/en-us/azure/active-directory/develop/v1-protocols-oauth-code

This parameter should be include in the login URL.