Fix uses of UserID x Email
FIXME: How to fix `GetJwtSession`? 1) Does the token possibly contain the selected user-id-claim?
2) How do I know here the value(s) of user-id-claim, which is only defined for the OIDC provider?

If it is us who created the JWT token then it has anything we need. But is it?
We could add a provider method to extract the ID claim from a `map[string]interface{}`
defaulting to email?