Align AES 16,24,32 CookieSecret requirements
CookieSecret is only required to be a valid AES length if
used in encryption. This is now only the case for Cookie Stores.

Redis uses its own dynamic 32 byte secret for each ticket for
AES-GCM encryption in the store. So it never needs to run in minimal
mode and strip the tokens from the session anymore.

In Redis legacy fallback mode, it will only attempt a legacy session
load if the CookieSecret is a valid length -- since then it is required
to decode the legacy session style that had embedded AES-CFB encryption.