Add automatic cookie_refresh mode and a grace period
Currently when a OIDC tokens expire the session cookie is not refreshed by default resulting in proxy providing a useless stale token.
The cookie_refresh mode that is based on TTL is largely useless as that TTL needs to be synchronized with settings in OIDC provider, which also makes it largely unusable.

Thus, an additional automatic cookie refresh mode is added as follows:

    cookie_refresh = -1
    cookie_refresh_grace_pcnt = 85

When cookie_refresh is set to -1 (default is still 0 for disabled to preserve backwards compatibility) every cookie refresh will undergo a check through the provider to see if the cookie needs to be refreshed based on token expiry. The refresh grace percentage (default to 100%) is then used to determine when to set the expiry. The grace percentage of 100% means that the token/cookie will be refreshed at 100% of the expiration period. As in the example above setting it to any other number X between 1 - 99 will refresh the token after X% of the token lifetime has expired.

As an additional piece of logic, the expiry time for OIDC provider is now set to the earliest of ID Token and Access Token expiration.

A bonus fix - when selecting a provider, the choice is now printed and a warning is issued if provider name is not recognized (I got burned by this personally not realizing "openidc" is not "oidc" and there would be a default to "google").

Lastly, architecturally, the APIs for providers now carry ProxyState as a parameter, granting providers access to relevant common configuration. I only threw there what I needed for this.

fixes #856