Properly skip preflight checks in auth-only (no proxy) mode
Right now, all requests that give true for IsAllowedRequest are passed
to upstream, even if the request is for "/oauth2/auth". This is fine when
oauth2-proxy is used in reverse proxy mode, but gives unexpected 404 errors
for auto-only mode.

This PR fixes this by properly handing /oauth2/auth in SkipAuthProxy by
not passing it to upstream and instead returning 202.