Build scratch images
We are seeing quite a bit of vulnerabilities in the base images
from time to time. And since this is a staticly linked Go binary
we have the possiblity to run it from a scratch image reducing the
attack surface.

We need to copy over the public root certificates in order for SSL calls
to be supported.

We are now building the following images:
   latest
   latest-scratch
   VERSION
   VERSION-scratch

All of these images are build for the supported platforms.

Please note, this remove the explicit tagging of images with the
platform architecture like latest-arm64.

If one needs an explicit image you can pull with --platform=arm64