Return CSRF token as cookie (#2573)
Introduce helpers that will wrap the CSRF token in a cookie. Set the
cookie along with the session cookie, when the callback function
is called. Expose the following flags to configure CSRF token cookie
attributes:

* --csrftoken-cookie-name: the name of the CSRF token cookie (default:
  _oauth2_proxy_csrftoken). If set to an empty string, the functionality
  is disabled.
* --csrftoken-cookie-domain: the domain(s) of the cookie (optional, default "")
* --csrftoken-cookie-path: the path of the cookie (optional, default: /)
* --csrftoken-cookie-expire: expiration of the cookie (default: 168 hours)
* --csrftoken-cookie-secure: set secure (HTTPS) cookie (default: true)
* --csrftoken-cookie-httponly: set HTTPOnly for cookie (default: false)
* --csrftoken-cookie-samesite: set SameSite for cookie (default: strict)

Note that a CSRF cookie should be secure but not HttpOnly, since clients
are expected to retrieve the value of the token to send in subsequent
requests.

Set SameSite to strict by default.

Co-authored-by: Athina Plaskasoviti <athina.plaskasoviti@hpe.com>
Co-authored-by: Konstantinos Lolos <konstantinos.lolos@hpe.com>