Return CSRF token as custom header (#2573)
Expose flag --csrftoken-response-header (string) to allow returning the
session CSRF token as a custom header. Default header name is X-CSRF-Token,
which the proxy adds to every request if the session has a CSRF token.

If --csrftoken-response-header is set to an empty string, the functionality
is disabled.

**Note** that --csrftoken-response-header is not available when the proxy
runs using alpha configuration. To enable it with alpha configuration,
add the necessary entry in the alpha config in the injectResponseHeaders
field.

Co-authored-by: Athina Plaskasoviti <athina.plaskasoviti@hpe.com>
Co-authored-by: Konstantinos Lolos <konstantinos.lolos@hpe.com>